requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
azure-core==1.16.0
azure-storage-blob==12.8.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
greenlet==1.1.0
idna==3.2
importlib-metadata==4.6.1
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msrest==0.6.21
multidict==5.1.0
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.2
py4j==0.10.9
pycparser==2.20
pycryptodome==3.10.1
PyMySQL==1.0.2
pyodbc==4.0.31
pyspark==3.1.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
SQLAlchemy==1.4.21
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0
